# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Monitoreo del hogar",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "autenticación"
        ],
        "summary": "Login",
        "description": "Abre una sesión y la entrega en la cookie `sesion`.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Demasiados intentos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "autenticación"
        ],
        "summary": "Logout",
        "description": "Cierra la sesión actual y borra la cookie.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/yo": {
      "get": {
        "tags": [
          "autenticación"
        ],
        "summary": "Yo",
        "description": "El usuario de la sesión y sus casas, con su rol en cada una.",
        "operationId": "yo_api_v1_auth_yo_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Yo"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/cambiar-clave": {
      "post": {
        "tags": [
          "autenticación"
        ],
        "summary": "Cambiar Clave",
        "description": "Cambia la clave y cierra las demás sesiones del usuario (§12.2).",
        "operationId": "cambiar_clave_api_v1_auth_cambiar_clave_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioClave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sesiones": {
      "get": {
        "tags": [
          "autenticación"
        ],
        "summary": "Sesiones",
        "description": "Las sesiones abiertas del usuario, de la más reciente a la más antigua.",
        "operationId": "sesiones_api_v1_auth_sesiones_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SesionAbierta"
                  },
                  "type": "array",
                  "title": "Response Sesiones Api V1 Auth Sesiones Get"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sesiones/{sesion_id}": {
      "delete": {
        "tags": [
          "autenticación"
        ],
        "summary": "Cerrar Sesion",
        "description": "Cierra una de las sesiones del usuario (por ejemplo, la de un celular perdido).",
        "operationId": "cerrar_sesion_api_v1_auth_sesiones__sesion_id__delete",
        "parameters": [
          {
            "name": "sesion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sesion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/casas/{casa_id}/ajustes": {
      "get": {
        "tags": [
          "casas"
        ],
        "summary": "Ver",
        "operationId": "ver_api_v1_casas__casa_id__ajustes_get",
        "parameters": [
          {
            "name": "casa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Casa Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ajustes"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "casas"
        ],
        "summary": "Cambiar",
        "description": "Cambia solo los campos enviados; un null se ignora.",
        "operationId": "cambiar_api_v1_casas__casa_id__ajustes_patch",
        "parameters": [
          {
            "name": "casa_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Casa Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambioAjustes"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ajustes"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso u origen inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ajustes": {
        "properties": {
          "recordatorio_min": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Recordatorio Min"
          },
          "recordatorio_conexion_min": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Recordatorio Conexion Min"
          },
          "minutos_central_caida": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 1.0,
            "title": "Minutos Central Caida"
          },
          "avisar_nodo_sin_conexion": {
            "type": "boolean",
            "title": "Avisar Nodo Sin Conexion"
          },
          "avisar_resueltas": {
            "type": "boolean",
            "title": "Avisar Resueltas"
          }
        },
        "type": "object",
        "required": [
          "recordatorio_min",
          "recordatorio_conexion_min",
          "minutos_central_caida",
          "avisar_nodo_sin_conexion",
          "avisar_resueltas"
        ],
        "title": "Ajustes"
      },
      "CambioAjustes": {
        "properties": {
          "recordatorio_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordatorio Min"
          },
          "recordatorio_conexion_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordatorio Conexion Min"
          },
          "minutos_central_caida": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutos Central Caida"
          },
          "avisar_nodo_sin_conexion": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avisar Nodo Sin Conexion"
          },
          "avisar_resueltas": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avisar Resueltas"
          }
        },
        "type": "object",
        "title": "CambioAjustes",
        "description": "Solo cambian los campos que se envían."
      },
...
      "CambioClave": {
        "properties": {
          "clave_actual": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Clave Actual"
          },
          "clave_nueva": {
            "type": "string",
            "maxLength": 128,
            "minLength": 10,
            "title": "Clave Nueva"
          }
        },
        "type": "object",
        "required": [
          "clave_actual",
          "clave_nueva"
        ],
        "title": "CambioClave"
      },
...
      "CasaDelUsuario": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "rol": {
            "type": "string",
            "enum": [
              "admin",
              "cuidador"
            ],
            "title": "Rol"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "rol"
        ],
        "title": "CasaDelUsuario"
      },
...
      "DetalleError": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "campos": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campos"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "mensaje"
        ],
        "title": "DetalleError"
      },
      "Error": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/DetalleError"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "Error",
        "description": "{\"detail\": {\"codigo\": ..., \"mensaje\": ...}}"
      },
      "Login": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "clave": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Clave"
          }
        },
        "type": "object",
        "required": [
          "email",
          "clave"
        ],
        "title": "Login"
      },
...
      "SesionAbierta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "creada_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creada En"
          },
          "ultimo_uso_en": {
            "type": "string",
            "format": "date-time",
            "title": "Ultimo Uso En"
          },
          "expira_en": {
            "type": "string",
            "format": "date-time",
            "title": "Expira En"
          },
          "actual": {
            "type": "boolean",
            "title": "Actual"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_agent",
          "ip",
          "creada_en",
          "ultimo_uso_en",
          "expira_en",
          "actual"
        ],
        "title": "SesionAbierta"
      },
      "Usuario": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "es_superadmin": {
            "type": "boolean",
            "title": "Es Superadmin"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "nombre",
          "es_superadmin"
        ],
        "title": "Usuario"
      },
      "Yo": {
        "properties": {
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "casas": {
            "items": {
              "$ref": "#/components/schemas/CasaDelUsuario"
            },
            "type": "array",
            "title": "Casas"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "casas"
        ],
        "title": "Yo"
      }
    }
  }
}